# api/openapi.yaml
openapi: 3.1.0
info:
  title: SimpWF Workflow Engine
  version: 1.0.0
  description: |
    Durable PostgreSQL-backed workflow engine. Definitions are immutable;
    new versions reference the previous version. Instances run asynchronously
    and expose status, context, input, and node-level debugging APIs.
servers:
  - url: /
...
paths:
...
  /v1/node/definition:
    post:
      tags: [Node Definitions]
      operationId: createNodeDefinition
      security:
        - ApiKeyAuth: []
      summary: Create an immutable node definition (or its next version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NodeDefinitionRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeDefinition" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    get:
      tags: [Node Definitions]
      operationId: listNodeDefinitions
      security:
        - ApiKeyAuth: []
      summary: List node definitions
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/Name"
        - $ref: "#/components/parameters/LineageID"
        - $ref: "#/components/parameters/Version"
        - $ref: "#/components/parameters/LatestOnly"
        - $ref: "#/components/parameters/NodeType"
      responses:
        "200":
          description: List of node definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDefinitionList"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/node/definition/{id}:
    get:
      tags: [Node Definitions]
      operationId: getNodeDefinition
      security:
        - ApiKeyAuth: []
      summary: Get a node definition by id
      parameters:
        - $ref: "#/components/parameters/DefinitionID"
      responses:
        "200":
          description: Node definition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeDefinition" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [Node Definitions]
      operationId: deleteNodeDefinition
      security:
        - ApiKeyAuth: []
      summary: Delete a node definition (fails if referenced)
      parameters:
        - $ref: "#/components/parameters/DefinitionID"
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Conflict" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/workflow/definition:
    post:
      tags: [Workflow Definitions]
      operationId: createWorkflowDefinition
      security:
        - ApiKeyAuth: []
      summary: Create an immutable workflow definition (or its next version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkflowDefinitionRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkflowDefinition" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    get:
      tags: [Workflow Definitions]
      operationId: listWorkflowDefinitions
      security:
        - ApiKeyAuth: []
      summary: List workflow definitions
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/ID"
        - $ref: "#/components/parameters/Name"
        - $ref: "#/components/parameters/LineageID"
        - $ref: "#/components/parameters/Version"
        - $ref: "#/components/parameters/LatestOnly"
      responses:
        "200":
          description: List of workflow definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinitionList"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/workflow/definition/{id}:
    get:
      tags: [Workflow Definitions]
      operationId: getWorkflowDefinition
      security:
        - ApiKeyAuth: []
      summary: Get a workflow definition by id
      parameters:
        - $ref: "#/components/parameters/DefinitionID"
      responses:
        "200":
          description: Workflow definition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkflowDefinition" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [Workflow Definitions]
      operationId: deleteWorkflowDefinition
      security:
        - ApiKeyAuth: []
      summary: Delete a workflow definition (fails if referenced by instances)
      parameters:
        - $ref: "#/components/parameters/DefinitionID"
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Conflict" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Token
      description: API token authentication. Required when auth is enabled; sent as X-Api-Token header.
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Order:
      name: order
      in: query
      required: false
      schema:
        type: string
        description: Sort field with optional leading "-" for descending (allowlisted)
    ID:
      name: id
      in: query
      required: false
      schema:
        type: array
        items: { type: string, format: uuid }
        maxItems: 100
        description: Repeatable id filter
      style: form
      explode: true
    Name:
      name: name
      in: query
      required: false
      schema: { type: string }
    LineageID:
      name: lineage_id
      in: query
      required: false
      schema: { type: string, format: uuid }
    Version:
      name: version
      in: query
      required: false
      schema: { type: integer, minimum: 1 }
    LatestOnly:
      name: latest_only
      in: query
      required: false
      schema: { type: boolean }
    NodeType:
      name: type
      in: query
      required: false
      schema: { type: string }
...
    DefinitionID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid query or malformed request
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: State or reference conflict
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Missing or invalid API token
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unprocessable:
      description: Invalid definition or input
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    NodeDefinitionRequest:
      type: object
      required: [name, type, content]
      properties:
        name: { type: string }
        type:
          type: string
          enum: [script, conditions, input, group, external_call]
        previous_version_id: { type: string, format: uuid, nullable: true }
        content: { type: object }
    NodeDefinition:
      type: object
      required: [id, name, version, lineage_id, type, content, created_by, updated_by, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        version: { type: integer }
        previous_version_id: { type: string, format: uuid, nullable: true }
        lineage_id: { type: string, format: uuid }
        type: { type: string }
        content: { type: object }
        created_by: { type: string, format: uuid }
        updated_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NodeDefinitionList:
      type: object
      required: [items, page, per_page, total, total_pages]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/NodeDefinition" } }
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
    WorkflowDefinitionRequest:
      type: object
      required: [name, content]
      properties:
        name: { type: string }
        previous_version_id: { type: string, format: uuid, nullable: true }
        content: { type: object }
    WorkflowDefinition:
      type: object
      required: [id, name, version, lineage_id, content, created_by, updated_by, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        version: { type: integer }
        previous_version_id: { type: string, format: uuid, nullable: true }
        lineage_id: { type: string, format: uuid }
        content: { type: object }
        created_by: { type: string, format: uuid }
        updated_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkflowDefinitionList:
      type: object
      required: [items, page, per_page, total, total_pages]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/WorkflowDefinition" } }
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
...